net[linux]: implements NetConnections() using lsof.
