+++ net/net_darwin.go
@@ -3,22 +3,12 @@
 package net
 
 import (
-	"fmt"
-	"net"
 	"os/exec"
 	"strconv"
 	"strings"
-	"syscall"
 
 	"github.com/shirou/gopsutil/common"
 )
-
-var constMap = map[string]int{
-	"TCP":  syscall.SOCK_STREAM,
-	"UDP":  syscall.SOCK_DGRAM,
-	"IPv4": syscall.AF_INET,
-	"IPv6": syscall.AF_INET6,
-}
 
 // example of netstat -idbn output on yosemite
 // Name  Mtu   Network       Address            Ipkts Ierrs     Ibytes    Opkts Oerrs     Obytes  Coll Drop
@@ -154,75 +144,4 @@ func NetConnections(kind string) ([]NetConnectionStat, error) {
 	}
 
 	return ret, nil
-}
-
-func parseNetLine(line string) (NetConnectionStat, error) {
-	f := strings.Fields(line)
-	if len(f) < 9 {
-		return NetConnectionStat{}, fmt.Errorf("wrong line,%s", line)
-	}
-
-	pid, err := strconv.Atoi(f[1])
-	if err != nil {
-		return NetConnectionStat{}, err
-	}
-	fd, err := strconv.Atoi(strings.Trim(f[3], "u"))
-	if err != nil {
-		return NetConnectionStat{}, fmt.Errorf("unknown fd, %s", f[3])
-	}
-	netFamily, ok := constMap[f[4]]
-	if !ok {
-		return NetConnectionStat{}, fmt.Errorf("unknown family, %s", f[4])
-	}
-	netType, ok := constMap[f[7]]
-	if !ok {
-		return NetConnectionStat{}, fmt.Errorf("unknown type, %s", f[7])
-	}
-
-	laddr, raddr, err := parseNetAddr(f[8])
-	if err != nil {
-		return NetConnectionStat{}, fmt.Errorf("failed to parse netaddr, %s", f[8])
-	}
-
-	n := NetConnectionStat{
-		Fd:     uint32(fd),
-		Family: uint32(netFamily),
-		Type:   uint32(netType),
-		Laddr:  laddr,
-		Raddr:  raddr,
-		Pid:    int32(pid),
-	}
-	if len(f) == 10 {
-		n.Status = strings.Trim(f[9], "()")
-	}
-
-	return n, nil
-}
-
-func parseNetAddr(line string) (laddr Addr, raddr Addr, err error) {
-	parse := func(l string) (Addr, error) {
-		host, port, err := net.SplitHostPort(l)
-		if err != nil {
-			return Addr{}, fmt.Errorf("wrong addr, %s", l)
-		}
-		lport, err := strconv.Atoi(port)
-		if err != nil {
-			return Addr{}, err
-		}
-		return Addr{IP: host, Port: uint32(lport)}, nil
-	}
-
-	addrs := strings.Split(line, "->")
-	if len(addrs) == 0 {
-		return laddr, raddr, fmt.Errorf("wrong netaddr, %s", line)
-	}
-	laddr, err = parse(addrs[0])
-	if len(addrs) == 2 { // remote addr exists
-		raddr, err = parse(addrs[1])
-		if err != nil {
-			return laddr, raddr, err
-		}
-	}
-
-	return laddr, raddr, err
 }
