+++ net/net_linux.go
@@ -93,5 +93,53 @@ func NetIOCounters(pernic bool) ([]NetIOCountersStat, error) {
 func NetConnections(kind string) ([]NetConnectionStat, error) {
 	var ret []NetConnectionStat
 
-	return ret, common.NotImplementedError
+	args := []string{"-i"}
+	switch strings.ToLower(kind) {
+	default:
+		fallthrough
+	case "":
+		fallthrough
+	case "all":
+		fallthrough
+	case "inet":
+		args = append(args, "tcp")
+	case "inet4":
+		args = append(args, "4")
+	case "inet6":
+		args = append(args, "6")
+	case "tcp":
+		args = append(args, "tcp")
+	case "tcp4":
+		args = append(args, "4tcp")
+	case "tcp6":
+		args = append(args, "6tcp")
+	case "udp":
+		args = append(args, "udp")
+	case "udp4":
+		args = append(args, "6udp")
+	case "udp6":
+		args = append(args, "6udp")
+	case "unix":
+		return ret, common.NotImplementedError
+	}
+
+	// we can not use -F filter to get all of required information at once.
+	r, err := common.CallLsof(invoke, 0, args...)
+	if err != nil {
+		return nil, err
+	}
+	for _, rr := range r {
+		if strings.HasPrefix(rr, "COMMAND") {
+			continue
+		}
+		n, err := parseNetLine(rr)
+		if err != nil {
+			// fmt.Println(err) // TODO: should debug print?
+			continue
+		}
+
+		ret = append(ret, n)
+	}
+
+	return ret, nil
 }
